requirements.txt
runpod
ddgs
dnspython